Testing different heuristic functions & diagonal movement cost
Looking to optimize cost for diagonals to prefer straight path movement.

+++ A_star/astar+bin/include/Astar.hpp
@@ -5,13 +5,13 @@
 
 struct Node {
     int x, y;
-    int g, h, f;
+    float g, h, f;
     Node* parent;
     bool isObstacle;
     bool start;
     bool exit;
 
-    Node(int x, int y) : x(x), y(y), g(0), h(0), f(0), parent(nullptr), isObstacle(false), start(false), exit(false) {}
+    Node(int x, int y) : x(x), y(y), g(0.0), h(0.0), f(0.0), parent(nullptr), isObstacle(false), start(false), exit(false) {}
 };
 
 class Astar {
@@ -26,7 +26,7 @@ class Astar {
         void initializeGrid();
         void cleanupGrid();
         bool isValid(int x, int y);
-        int heuristic(int x1, int y1, int x2, int y2);
+        float heuristic(int x1, int y1, int x2, int y2);
         std::vector<Node*> getNeighbors(Node* node);
         Node* findLowestF(std::vector<Node*>& openSet);
         std::vector<Node*> reconstructPath(Node* current);
